Apparently fileRename() is called quite a bit with bogus filenames.  So, only unlink if it is completely safe.

+++ src/asf/fileUtil.c
@@ -549,7 +549,8 @@ void fileRename(const char *src, const char *dst)
 {
     // FIXME
     // asfSystem("mv %s %s", src, dst);
-    unlink(dst);
+    if (fileExists(src) && fileExists(dst))
+      unlink(dst);
     rename(src, dst);
 }
